testing delay and message loss

+++ lab1_cs7610/prj1_tm.cpp
@@ -366,7 +366,7 @@ int main(int argc, char *argv[])
     priority_queue <Mesg_pq, vector<Mesg_pq>, CompareMessage> final_mesg_q;
     uint32_t loss_pid;
     if(argc == 3)
-        loss_pid = atoi(argv[2]) ;
+        loss_pid = (uint32_t ) atoi(argv[2]) ;
 
     FD_ZERO(&writefds);    // clear the write and temp sets
     FD_ZERO(&readfds);
